Discord_API/requirements.txt
# ===================================================================
# Discord и Asyncio (Core & Async)
# ===================================================================
# Основная библиотека для работы с Discord API
discord.py

# Высокопроизводительный цикл событий для asyncio
uvloop

# ===================================================================
# Взаимодействие с Backend и внешними API
# ===================================================================
# Асинхронный HTTP-клиент для запросов к вашему основному серверу
aiohttp
websockets

# Библиотека для умных повторных попыток при сбоях API
tenacity

# Клиент для RabbitMQ (если бот обменивается сообщениями с сервером)
aio-pika

# ===================================================================
# Работа с данными (Data & Databases)
# ===================================================================
# ORM и миграции для прямого доступа к БД (например, за данными пользователей)
SQLAlchemy
alembic
Mako

# Асинхронный драйвер для PostgreSQL
asyncpg
psycopg2-binary # Может быть нужен для alembic или других синхронных утилит

# Клиенты для Redis (кэширование, временные данные)
redis
aioredis

# Валидация данных, полученных от API сервера
Pydantic

# Быстрая бинарная сериализация (для Redis или RabbitMQ)
msgpack

# ===================================================================
# Утилиты и Логирование (Utilities & Logging)
# ===================================================================
# Загрузка переменных окружения из .env файлов
python-dotenv

# Библиотеки для логирования
loguru
colorlog

# Мониторинг для Prometheus (кастомные метрики бота)
prometheus-client

# Получение системной информации (для отладочных команд)
psutil